synapse-backend/services/uploads/requirements.txt
boto3==1.28.64
minio==7.1.17
PyJWT==2.8.0
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
asyncpg==0.28.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
httpx==0.25.1
redis==4.6.0
python-magic==0.4.27
aiofiles==23.2.1
pillow==10.1.0
boto3==1.28.64
minio==7.1.17
